backend/requirements.txt
fastapi
uvicorn[standard]
numpy
scipy
xarray
netCDF4
pandas
scikit-learn
torch>=2.2.0
torch-geometric>=2.5.0
torch-scatter
torch-sparse
copernicusmarine
earthengine-api
python-dotenv
requests
tqdm
matplotlib
